+++ src/modules/Uploader/index.js
@@ -1,0 +1,4 @@
+
+import { Uploader } from "@prisma-cms/ui";
+
+export default Uploader;